refactor: modifications and improvements * refactor test files * add tests to investigate plot module bugs
